refactor: extend UIViewController to present UIAlertController
