+++ CookAndShare/Controller/Share/ChatRoomViewController.swift
@@ -187,12 +187,7 @@ class ChatRoomViewController: UIViewController {
 
     @objc func blockUser() {
         guard let friend = friend else { return }
-        let alert = UIAlertController(
-            title: "封鎖\(friend.name)？",
-            message: "你將不會看到他的貼文、個人檔案或來自他的訊息。你封鎖用戶時，對方不會收到通知。",
-            preferredStyle: .actionSheet
-        )
-        let confirmAction = UIAlertAction(title: "確定封鎖", style: .destructive) { [weak self] _ in
+        let handler: AlertActionHandler = { [weak self] _ in
             guard let self = self else { return }
             let myRef = FirestoreEndpoint.users.collectionRef.document(Constant.getUserId())
             self.firestoreManager.arrayUnionString(
@@ -202,22 +197,14 @@ class ChatRoomViewController: UIViewController {
             )
             self.navigationController?.popToRootViewController(animated: true)
         }
-        let cancelAction = UIAlertAction(title: "取消", style: .cancel)
-        alert.addAction(confirmAction)
-        alert.addAction(cancelAction)
 
-        if let popoverController = alert.popoverPresentationController {
-            popoverController.sourceView = self.view
-            popoverController.sourceRect = CGRect(
-                x: self.view.bounds.midX,
-                y: self.view.bounds.midY,
-                width: 0,
-                height: 0
-            )
-            popoverController.permittedArrowDirections = []
-        }
-
-        present(alert, animated: true)
+        presentAlertWith(
+            alertTitle: "封鎖\(friend.name)？",
+            alertMessage: "你將不會看到他的貼文、個人檔案或來自他的訊息。你封鎖用戶時，對方不會收到通知。",
+            confirmTitle: "確定封鎖",
+            cancelTitle: "取消",
+            handler: handler
+        )
     }
 
     func uploadMessage(contentType: ContentType, content: String, duration: Double = 0) {
@@ -319,15 +306,17 @@ class ChatRoomViewController: UIViewController {
     }
 
     func alertSendingLocation() {
-        let alert = UIAlertController(title: "是否傳送目前位置？", message: nil, preferredStyle: .alert)
-        let okAction = UIAlertAction(title: Constant.confirm, style: .default) { _ in
+        let handler: AlertActionHandler = { _ in
             let locationString = "\(self.location.coordinate.latitude),\(self.location.coordinate.longitude)"
             self.uploadMessage(contentType: .location, content: locationString)
         }
-        let cancelAction = UIAlertAction(title: Constant.cancel, style: .cancel, handler: nil)
-        alert.addAction(okAction)
-        alert.addAction(cancelAction)
-        present(alert, animated: true, completion: nil)
+        presentAlertWith(
+            alertTitle: "是否傳送目前位置？",
+            alertMessage: nil,
+            confirmTitle: Constant.confirm,
+            cancelTitle: Constant.cancel,
+            handler: handler
+        )
     }
 
     func alertLocationAccessNeeded() {
