+++ CookAndShare/Controller/Share/ShareViewController.swift
@@ -13,7 +13,6 @@ import ESPullToRefresh
 import SPAlert
 
 class ShareViewController: UIViewController {
-    typealias AlertActionHandler = (UIAlertAction) -> Void
     var shareId = ""
     var fromPublicVC = false
     private let firestoreManager = FirestoreManager.shared
@@ -187,31 +186,6 @@ class ShareViewController: UIViewController {
         self.shares = tempShares.sorted { $0.postTime.seconds > $1.postTime.seconds }
     }
 
-    private func presentAlertController(alertTitle: String, alertMessage: String, confirmTitle: String, cancelTitle: String, handler: ((UIAlertAction) -> Void)?) {
-        let alert = UIAlertController(
-            title: alertTitle,
-            message: alertMessage,
-            preferredStyle: .actionSheet
-        )
-        let confirmAction = UIAlertAction(title: confirmTitle, style: .destructive, handler: handler)
-        let cancelAction = UIAlertAction(title: cancelTitle, style: .cancel)
-        alert.addAction(confirmAction)
-        alert.addAction(cancelAction)
-
-        if let popoverController = alert.popoverPresentationController {
-            popoverController.sourceView = self.view
-            popoverController.sourceRect = CGRect(
-                x: self.view.bounds.midX,
-                y: self.view.bounds.midY,
-                width: 0,
-                height: 0
-            )
-            popoverController.permittedArrowDirections = []
-        }
-
-        present(alert, animated: true)
-    }
-
     // MARK: - Action
     @objc func addShare() {
         if Auth.auth().currentUser == nil {
@@ -298,7 +272,7 @@ extension ShareViewController: ShareCellDelegate {
             )
         }
 
-        presentAlertController(
+        presentAlertWith(
             alertTitle: "確定刪除此貼文？",
             alertMessage: "此動作將無法回復。",
             confirmTitle: "確定刪除",
@@ -330,7 +304,7 @@ extension ShareViewController: ShareCellDelegate {
             )
         }
 
-        presentAlertController(
+        presentAlertWith(
             alertTitle: "封鎖\(user.name)？",
             alertMessage: "你將不會看到他的貼文、個人檔案或來自他的訊息。你封鎖用戶時，對方不會收到通知。",
             confirmTitle: "確定封鎖",
@@ -353,7 +327,7 @@ extension ShareViewController: ShareCellDelegate {
             SPAlert.present(message: "謝謝你告知我們，我們會在未來減少顯示這類內容", haptic: .success)
         }
 
-        presentAlertController(
+        presentAlertWith(
             alertTitle: "檢舉這則貼文？",
             alertMessage: "你的檢舉將會匿名。",
             confirmTitle: "確定檢舉",
